[HAKWULAR-844] some broken itests
Initially I was putting tests in a separate module so that I could test the bus i
as a JBoss module. That is when I discovered CDI injection does not work with
JBoss modules, at least with what I have done. At least for now the tests can
live in the hawkular-bus-common module and be executed in the test phase which
makes for easier and faster test execution. The Bus class is a just a prototype.