[HAWKULAR-844] work around for class name header
We set a header on each message for the class name which is resulting in the
following exception with Artemis,

javax.ejb.EJBException: javax.jms.JMSRuntimeException: AMQ129012: The property name 'org.hawkular.bus.common.BasicMessage.className' is not a valid java identifier.
    Caused by: javax.jms.JMSRuntimeException: AMQ129012: The property name 'org.hawkular.bus.common.BasicMessage.className' is not a valid java identifier."}}

At least as a temporary work around I have changed the header name to a
recognized, supported name.