[HAWKULAR-844] a couple minor fixes for module deployment
We need a dependency on the jackson annotation module. I discovered that during
manual testing. I also modified the build slightly to include the assembly id
in the zip artifact name because the main hawkular build expects this and will
not lay down the module properly if the assembly id is not included. Lastly, I
have removed constructors that were used specifically with ActiveMQ.