Switch to copy-module-artifacts="true" in server-provisioning.xml
This is necessary because with copy-module-artifacts="false", the
artifacts are looked for in local maven repo where they are not
available before the install phase, which is after both test and
integration tests-phases. If this sometimes works with
copy-module-artifacts="false", it caused by either consuming the
artifacts in the local repo from the previous builds or pulling
SNAPSHOTs from the SNAPSHOT maven repository.